Split backend chat live worker modules
