+++ apps/backend/src/chat/live/auth.ts
@@ -1,7 +1,7 @@
 import { createHmac, timingSafeEqual } from "node:crypto";
-import { HttpError } from "../errors";
-import type { BackendTraceCarrier } from "../observability/sentry";
-import { getBackendChatLiveAuthSecret } from "../secrets";
+import { HttpError } from "../../errors";
+import type { BackendTraceCarrier } from "../../observability/sentry";
+import { getBackendChatLiveAuthSecret } from "../../secrets";
 
 const CHAT_LIVE_AUTH_TOKEN_VERSION = 1;
 const CHAT_LIVE_AUTH_TTL_MS = 10 * 60 * 1000;
