+++ apps/backend/src/chat/tests/chat-live-diff.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import { diffAssistantContent } from "../liveDiff";
+import { diffAssistantContent } from "../live/diff";
 import type { ContentPart } from "../types";
 
 function createToolCallPart(params: Readonly<{
