+++ apps/backend/src/chat/tests/chat-live-resume.test.ts
@@ -4,7 +4,7 @@ import { PassThrough } from "node:stream";
 import {
   runLiveStreamWithDependencies,
   type ChatLiveStreamResult,
-} from "../live";
+} from "../live/index";
 import type { ChatComposerSuggestion } from "../composerSuggestions";
 import type { ChatRunSnapshot } from "../runs";
 import type { PersistedChatMessageItem } from "../store";
